preparing data and then label trips
In this project, we collected some data from mobilita dynamica application that related to coordinate of trips that persons used this application.In fact final purpose of project was dynamic label zone and determining purpose of trips. first of all we prepared data for using machine learning algorithm to make a model for next prediction. also we labeled the zones of Piedmont according to number of trips and attraction place in each zone .the tools that we used in this project are mongodb and python.
python 1: preparing data and then label trips according to nearest attraction place .